feat: seed.habit 데이터 추가 및 seeds 폴더 생성
