+++ package.json
@@ -5,9 +5,9 @@
   "main": "src/index.js",
   "scripts": {
     "start": "node app.js",
-    "dev": "nodemon app.js",
+    "dev": "CHOKIDAR_USEPOLLING=true nodemon app.js",
     "build": "npm install",
-    "seed": "node prisma/seed.js"
+    "seed": "node prisma/seeds/seed.js"
   },
   "dependencies": {
     "@prisma/client": "^6.3.0",
