+++ prisma/seeds/seed.reaction.js
@@ -1,0 +1,23 @@
+export const reactions = [
+  {
+    id: crypto.randomUUID(),
+    emoji: "😀",
+    counts: 10,
+    createdAt: new Date(),
+    updatedAt: new Date(),
+  },
+  {
+    id: crypto.randomUUID(),
+    emoji: "😁",
+    counts: 14,
+    createdAt: new Date(),
+    updatedAt: new Date(),
+  },
+  {
+    id: crypto.randomUUID(),
+    emoji: "🤣",
+    counts: 30,
+    createdAt: new Date(),
+    updatedAt: new Date(),
+  },
+];
